requirements.txt
google-generativeai==0.3.2
pvporcupine==3.0.1
pyaudio==0.2.13
SpeechRecognition==3.10.0
pyttsx3==2.90
python-dotenv==1.0.0
requests==2.31.0 